Move set_executable_permissions into fs module
